Fixed debugger quit not working when inside a task group
